uClibc: fix darwin build
uClibc's unifdef tool declares strlcpy(3) which is problematic on OS X. There
(and probably on other systems) strlcpy(3) is defined as a macro, re-declaring
it here will fail cause the macro expansion will generate an invalid function
name.
The declaration of strlcpy(3) is however useless cause it is never used in the
tool. Also upstream patch 2a021ae81c36f4281883a3195f7ce81504edf978 will remove
it. Using that patch here is overkill, cause it rewrites a lot of the
unifdef.c code.

Signed-off-by: Andreas Bießmann <andreas@biessmann.de>
Cc: Yann E. MORIN <yann.morin.1998@free.fr>